Solving issue #26 and refactoring code in chatbot.py
