+++ .gitignore
@@ -1,1 +1,3 @@
 stanford-corenlp-full-2018-10-05/
+config.py
+__pycache__
